Reverting sentaku'd properties
* Keeping everything as commented out for now so it can be reverted
* Have the idea of a _fundamental_ that doesn't need to be picked
  properly
